+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 
 4 September
 - Added converting from ```NSString``` to ```UIColor```
+- Added method to return a ```NSArray``` of ```UIColor```